Move version out of pkg
